+++ src/components/molecules/index.ts
@@ -1,14 +1,19 @@
+export {default as ChangeFiltersConfirmModal} from './ChangeFiltersConfirmModal';
+export {default as CheckedResourcesActionsMenu} from './CheckedResourcesActionsMenu';
 export {default as ClusterDiff} from './ClusterDiff';
+export {default as ContextMenu} from './ContextMenu';
 export {default as FilePatternList} from './FilePatternList';
 export {default as FormEditor} from './FormEditor';
 export {default as GraphView} from './GraphView';
 export {default as HelmChartModalConfirmWithNamespaceSelect} from './HelmChartModalConfirmWithNamespaceSelect';
+export {default as HotkeyLabel} from './HotkeyLabel';
+export {default as IconWithPopover} from './IconWithPopover';
 export {default as ImageDetails} from './ImageDetails';
+export {default as InputTags} from './InputTags';
 export {default as ModalConfirm} from './ModalConfirm';
 export {default as ModalConfirmWithNamespaceSelect} from './ModalConfirmWithNamespaceSelect';
-export {default as HotkeyLabel} from './HotkeyLabel';
-export {default as InputTags} from './InputTags';
 export {default as Monaco} from './Monaco';
+export {default as NotificationMarkdown} from './NotificationMarkdown';
 export {default as PreviewConfigurationDetails} from './PreviewConfigurationDetails';
 export {default as PreviewDropdown} from './PreviewDropdown';
 export {default as QuickActionCompare} from './QuickActionCompare';
