+++ src/components/organisms/NavigatorPane/CompareButton.tsx
@@ -9,7 +9,7 @@ import {TOOLTIP_DELAY} from '@constants/constants';
 import {compareToggled} from '@redux/compare';
 import {useAppDispatch} from '@redux/hooks';
 
-import WalkThrough from '@components/molecules/Walkthrough';
+import {Walkthrough} from '@molecules';
 
 export const CompareButton: React.FC<{children?: React.ReactNode}> = ({children}) => {
   const dispatch = useAppDispatch();
@@ -18,7 +18,7 @@ export const CompareButton: React.FC<{children?: React.ReactNode}> = ({children}
   };
 
   return (
-    <WalkThrough collection="release" step="compare">
+    <Walkthrough collection="release" step="compare">
       <Tooltip mouseEnterDelay={TOOLTIP_DELAY} title="Compare resources" placement="bottom">
         <Button
           onClick={onClickClusterComparison}
@@ -31,6 +31,6 @@ export const CompareButton: React.FC<{children?: React.ReactNode}> = ({children}
           {children}
         </Button>
       </Tooltip>
-    </WalkThrough>
+    </Walkthrough>
   );
 };
