+++ src/navsections/ClusterDiffSectionBlueprint/ResourceMatchNameDisplay.tsx
@@ -21,7 +21,7 @@ import {isKustomizationResource} from '@redux/services/kustomize';
 import {applyResource} from '@redux/thunks/applyResource';
 import {updateResource} from '@redux/thunks/updateResource';
 
-import ModalConfirmWithNamespaceSelect from '@components/molecules/ModalConfirmWithNamespaceSelect';
+import {ModalConfirmWithNamespaceSelect} from '@molecules';
 
 import {
   diffLocalToClusterResources,
