+++ src/navsections/HelmChartSectionBlueprint/HelmChartContextMenuWrapper.tsx
@@ -17,7 +17,7 @@ import {isInPreviewModeSelector} from '@redux/selectors';
 import {getAbsoluteFilePath} from '@redux/services/fileEntry';
 import {setRootFolder} from '@redux/thunks/setRootFolder';
 
-import ContextMenu from '@molecules/ContextMenu';
+import {ContextMenu} from '@molecules';
 
 import {useCreate, useDuplicate, useFilterByFileOrFolder, useProcessing, useRename} from '@hooks/fileTreeHooks';
